Update 4
Added messageids to the users and userIds to the messages.  Added linking in the schema between messages and users and requiring that they both require each others id.  Added a filter to the user resolver so that it filters through the list of messages and only returns a list of messages with the matching user id.  Added a third message that is linked to the second user so that user will display two messages when queried