Update 10
Added createToken function that creates a hashed token and signs it as well as verifies the token's secret with the server and assigns an expiry time of 30 minutes to the token.  Added signin method to the resolvers to handle user signin and validate if user and password exist in the database.  Added signin mutation to the schema to require a login and password and return a Token type.  Added password validation method to the user prototype.  The function uses bcrypt to compare input password with the stored password.