Update 12
Added admin role to users.  Added protection resolvers to check if a user is authenticated, check if user is admin and check if message creater is message owner. Users can only create  if they are authenticated and delete if they are authenticated and are the message owner.  A user can only delete users if they have the role ADMIN.  Added admin role to the schema and models.  Fixed bug with bcrypt where password hash wasn't being created and compared properly.