Update 19
Added dataloader package to handle batching requests to the gql server.  Added batchUsers function that maps input keys to user ids.  Added userLoader that is a new DataLoader object containing a cache and handling storage of keys and models.  Added dataloader to the server context for use in the resolvers.  Updated Message resolver to use the dataloader load function to get the required users from the db.